Fix remaining toolbar review findings
